requirements.txt
pandas == 1.5.0
python-dotenv == 1.0.0
python-binance == 1.0.17
matplotlib == 3.7.1
TA-Lib